requirements.txt
Django>=3.2
djangorestframework>=3.12.4
requests>=2.25.1
djangorestframework-simplejwt>=4.8.0
psycopg2>=2.8.6
